+++ launching/src/main/java/org/jboss/tools/ssp/eclipse/jdt/launching/StandardVMType.java
@@ -34,10 +34,10 @@ import org.jboss.tools.ssp.eclipse.jdt.core.JavaCoreConstants;
 import org.jboss.tools.ssp.eclipse.jdt.internal.launching.LibraryInfo;
 import org.jboss.tools.ssp.eclipse.jdt.internal.launching.StandardVM;
 import org.jboss.tools.ssp.eclipse.osgi.util.NLS;
+import org.jboss.tools.ssp.internal.launching.util.OSUtils;
 import org.jboss.tools.ssp.launching.LaunchingCore;
 import org.jboss.tools.ssp.launching.LaunchingSupportUtility;
 import org.jboss.tools.ssp.launching.LibraryInfoCache;
-import org.jboss.tools.ssp.launching.util.OSUtils;
 
 
 /**
