+++ launching/src/main/java/org/jboss/tools/ssp/internal/launching/util/ExecUtil.java
@@ -6,7 +6,7 @@
  * 
  * Contributors: Red Hat, Inc.
  ******************************************************************************/
-package org.jboss.tools.ssp.launching.util;
+package org.jboss.tools.ssp.internal.launching.util;
 
 import java.io.File;
 import java.io.IOException;
