+++ launching/src/main/java/org/jboss/tools/ssp/launching/LaunchingSupportUtility.java
@@ -39,8 +39,8 @@ import org.jboss.tools.ssp.eclipse.jdt.launching.IVMInstallChangedListener;
 import org.jboss.tools.ssp.eclipse.jdt.launching.IVMRunner;
 import org.jboss.tools.ssp.eclipse.jdt.launching.VMRunnerConfiguration;
 import org.jboss.tools.ssp.eclipse.osgi.util.NLS;
-import org.jboss.tools.ssp.launching.util.ExecUtil;
-import org.jboss.tools.ssp.launching.util.OSUtils;
+import org.jboss.tools.ssp.internal.launching.util.ExecUtil;
+import org.jboss.tools.ssp.internal.launching.util.OSUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
