+++ launching/src/test/java/org/jboss/tools/ssp/launching/test/LaunchingJarTest.java
@@ -21,8 +21,8 @@ import org.jboss.tools.ssp.eclipse.core.runtime.NullProgressMonitor;
 import org.jboss.tools.ssp.eclipse.jdt.internal.launching.LibraryInfo;
 import org.jboss.tools.ssp.eclipse.jdt.launching.IVMInstall;
 import org.jboss.tools.ssp.eclipse.jdt.launching.StandardVMType;
+import org.jboss.tools.ssp.internal.launching.util.FileUtil;
 import org.jboss.tools.ssp.launching.LaunchingSupportUtility;
-import org.jboss.tools.ssp.launching.util.FileUtil;
 import org.junit.Test;
 
 public class LaunchingJarTest {
