+++ server-wildfly/META-INF/MANIFEST.MF
@@ -6,8 +6,7 @@ Bundle-Activator: org.jboss.tools.ssp.server.wildfly.impl.Activator
 Automatic-Module-Name: org.jboss.tools.ssp.server.wildfly
 Bundle-Version: 0.0.9.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Import-Package: org.osgi.framework,
- org.eclipse.lsp4j,
+Import-Package: org.eclipse.lsp4j,
  org.eclipse.lsp4j.generator,
  org.eclipse.lsp4j.jsonrpc,
  org.eclipse.lsp4j.jsonrpc.json,
@@ -28,9 +27,9 @@ Import-Package: org.osgi.framework,
  org.jboss.tools.ssp.eclipse.jdt.internal.launching,
  org.jboss.tools.ssp.eclipse.jdt.launching,
  org.jboss.tools.ssp.eclipse.osgi.util,
+ org.jboss.tools.ssp.internal.launching.util,
  org.jboss.tools.ssp.launching,
  org.jboss.tools.ssp.launching.test,
- org.jboss.tools.ssp.launching.util,
  org.jboss.tools.ssp.server,
  org.jboss.tools.ssp.server.core.internal,
  org.jboss.tools.ssp.server.discovery,
@@ -40,4 +39,5 @@ Import-Package: org.osgi.framework,
  org.jboss.tools.ssp.server.model.internal,
  org.jboss.tools.ssp.server.spi.discovery,
  org.jboss.tools.ssp.server.spi.model,
- org.jboss.tools.ssp.server.spi.servertype
+ org.jboss.tools.ssp.server.spi.servertype,
+ org.osgi.framework
