+++ server-wildfly/src/main/java/org/jboss/tools/ssp/server/wildfly/beans/impl/IJBossServerResourceConstants.java
@@ -1,6 +1,6 @@
 package org.jboss.tools.ssp.server.wildfly.beans.impl;
 
-import org.jboss.tools.ssp.launching.util.FileUtil;
+import org.jboss.tools.ssp.internal.launching.util.FileUtil;
 
 /**
  * Constants having to do with resources, folder paths,
