+++ server-wildfly/src/main/java/org/jboss/tools/ssp/server/wildfly/beans/impl/ServerBeanTypeUnknownAS71Product.java
@@ -14,7 +14,7 @@ import java.io.File;
 import java.nio.file.Path;
 import java.util.Properties;
 
-import org.jboss.tools.ssp.launching.util.FileUtil;
+import org.jboss.tools.ssp.internal.launching.util.FileUtil;
 import org.jboss.tools.ssp.server.wildfly.impl.util.JBossManifestUtility;
 
 public class ServerBeanTypeUnknownAS71Product extends JBossServerBeanType {
