+++ server/META-INF/MANIFEST.MF
@@ -6,8 +6,7 @@ Automatic-Module-Name: org.jboss.tools.ssp.server
 Bundle-Version: 0.0.9.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Activator: org.jboss.tools.ssp.server.ServerCoreActivator
-Import-Package: org.osgi.framework,
- org.eclipse.lsp4j,
+Import-Package: org.eclipse.lsp4j,
  org.eclipse.lsp4j.generator,
  org.eclipse.lsp4j.jsonrpc,
  org.eclipse.lsp4j.jsonrpc.json,
@@ -28,11 +27,12 @@ Import-Package: org.osgi.framework,
  org.jboss.tools.ssp.eclipse.jdt.internal.launching,
  org.jboss.tools.ssp.eclipse.jdt.launching,
  org.jboss.tools.ssp.eclipse.osgi.util,
+ org.jboss.tools.ssp.internal.launching.util,
  org.jboss.tools.ssp.launching,
- org.jboss.tools.ssp.launching.util,
  org.jboss.tools.ssp.server.spi.discovery,
  org.jboss.tools.ssp.server.spi.model,
- org.jboss.tools.ssp.server.spi.servertype
+ org.jboss.tools.ssp.server.spi.servertype,
+ org.osgi.framework
 Export-Package: org.jboss.tools.ssp.server,
  org.jboss.tools.ssp.server.core.internal,
  org.jboss.tools.ssp.server.discovery,
