+++ server/src/main/java/org/jboss/tools/ssp/server/discovery/serverbeans/util/ManifestUtility.java
@@ -19,7 +19,7 @@ import java.util.jar.Manifest;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
-import org.jboss.tools.ssp.launching.util.FileUtil;
+import org.jboss.tools.ssp.internal.launching.util.FileUtil;
 
 public class ManifestUtility {
